server/reveive_data/app/requirements.txt
paho-mqtt
numpy

